+++ CHANGELOG.md
@@ -43,6 +43,7 @@ log:
 - Bumped all dependencies
 - Enhanced the function signatures for the ref objects management 
 - Fixed a bug causing the jobs metrics pull to fail on ref.Kind=RefKindMergeRequest
+- Fixed a bug causing the environments to be garbage collected on every run
 - Fixed the error handling when comparing 2 refs which resulted into nil pointer dereferences
 - Fixed the pulling of merge-request based pipelines
 - Fixed unit tests on windows